change nota de credito

+++ src/main/java/com/paulfernandosr/possystembackend/contracts/infrastructure/adapter/output/row/VehicleSpecsRow.java
@@ -18,7 +18,7 @@ public class VehicleSpecsRow {
 
     private String engineCapacity;
     private String fuel;
-    private Integer cylinders;
+    private BigDecimal cylinders;
 
     private BigDecimal netWeight;
     private BigDecimal payload;
