+++ src/main/java/com/paulfernandosr/possystembackend/salev2/domain/port/output/ContractSunatDraftRepository.java
@@ -131,7 +131,7 @@ public interface ContractSunatDraftRepository {
         private String bodywork;
         private String engineCapacity;
         private String fuel;
-        private Integer cylinders;
+        private BigDecimal cylinders;
         private BigDecimal netWeight;
         private BigDecimal payload;
         private BigDecimal grossWeight;
