+++ src/main/java/com/paulfernandosr/possystembackend/salev2/domain/port/output/SaleV2SunatRepository.java
@@ -99,7 +99,7 @@ public interface SaleV2SunatRepository {
         private String bodywork;
         private String engineCapacity;
         private String fuel;
-        private Integer cylinders;
+        private BigDecimal cylinders;
         private BigDecimal netWeight;
         private BigDecimal payload;
         private BigDecimal grossWeight;
